feature: Add macro usage for autocomplete.
Elm notified me to this
[feature](https://github.com/not-elm/bevy-input-sequence/pull/10#issuecomment-2525100631).
I hope it will still work despite the code block being ignored. It's
ignored because my procedural macros use the `Modifiers` struct which is
located in the `keyseq` crate, so I can't actually have a compilable,
runnable test for the 'bevy' or 'winit' features of keyseq-macros.